# docs/lyc-api-openapi.yaml
openapi: 3.0.3
info:
  title: LYC 小宇宙学堂 API
  description: LYC 儿童启蒙学习平台后端 API，涵盖认证、内容管理、媒体播放、AI 分析、学习进度等模块。
  version: 1.0.0
  contact:
    name: LYC Team
servers:
  - url: http://127.0.0.1:3002
    description: 本地开发
  - url: https://api.lyc.example.com
    description: 生产环境
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 小程序端登录后获取的 JWT Token
    adminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: 管理员 API Key
  schemas:
...
    MediaFile:
      type: object
      properties:
        id:
          type: string
        originalName:
          type: string
        mimetype:
          type: string
        size:
          type: integer
        fileType:
          type: string
          enum: [audio, video]
        status:
          type: string
          enum: [ready, deleted]
        url:
          type: string
        createdAt:
          type: string
          format: date-time
    Progress:
      type: object
      properties:
        courseId:
          type: string
        currentSegmentIndex:
          type: integer
        lastPositionMs:
          type: integer
        playCount:
          type: integer
        completedSegments:
          type: array
          items:
            type: integer
        totalPlayMs:
          type: integer
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /api/v1/upload/single:
    post:
      tags: [文件上传]
      summary: 单文件上传
      description: 上传单个音视频文件，最大 200MB。
      security:
        - adminKey: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                courseId:
                  type: string
                fileType:
                  type: string
                  enum: [audio, video]
                  default: audio
      responses:
        "200":
          description: 上传成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  file:
                    $ref: "#/components/schemas/MediaFile"
  /api/v1/upload/init:
    post:
      tags: [文件上传]
      summary: 分片上传初始化
      description: 初始化分片上传会话，返回 sessionId 和 uploadId。
      security:
        - adminKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [filename, mimetype, totalChunks]
              properties:
                filename:
                  type: string
                mimetype:
                  type: string
                totalSize:
                  type: integer
                totalChunks:
                  type: integer
                courseId:
                  type: string
                fileType:
                  type: string
                  enum: [audio, video]
      responses:
        "200":
          description: 初始化成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId:
                    type: string
                  uploadId:
                    type: string
                  key:
                    type: string
                  chunkSize:
                    type: integer
                    example: 5242880
  /api/v1/upload/chunk:
    post:
      tags: [文件上传]
      summary: 上传分片
      security:
        - adminKey: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [sessionId, partNumber, chunk]
              properties:
                sessionId:
                  type: string
                partNumber:
                  type: integer
                chunk:
                  type: string
                  format: binary
      responses:
        "200":
          description: 上传成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  partNumber:
                    type: integer
                  etag:
                    type: string
                  receivedChunks:
                    type: integer
                  totalChunks:
                    type: integer
  /api/v1/upload/complete:
    post:
      tags: [文件上传]
      summary: 合并分片
      security:
        - adminKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sessionId, parts]
              properties:
                sessionId:
                  type: string
                parts:
                  type: array
                  items:
                    type: object
                    properties:
                      partNumber:
                        type: integer
                      etag:
                        type: string
      responses:
        "200":
          description: 合并成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  file:
                    $ref: "#/components/schemas/MediaFile"
  /api/v1/upload/abort:
    post:
      tags: [文件上传]
      summary: 取消分片上传
      security:
        - adminKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sessionId]
              properties:
                sessionId:
                  type: string
      responses:
        "200":
          description: 取消成功
  /api/v1/upload/{fileId}:
    delete:
      tags: [文件上传]
      summary: 删除文件
      security:
        - adminKey: []
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 删除成功

  # ========== 媒体播放 ==========
...
  /api/v1/progress/sync:
    post:
      tags: [学习进度]
      summary: 同步学习进度
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [courseId]
              properties:
                courseId:
                  type: string
                currentSegmentIndex:
                  type: integer
                lastPositionMs:
                  type: integer
                playCount:
                  type: integer
                completedSegments:
                  type: array
                  items:
                    type: integer
                totalPlayMs:
                  type: integer
      responses:
        "200":
          description: 同步成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  progress:
                    $ref: "#/components/schemas/Progress"
  /api/v1/progress/{courseId}:
    get:
      tags: [学习进度]
      summary: 获取单课进度
      security:
        - bearerAuth: []
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  progress:
                    $ref: "#/components/schemas/Progress"
  /api/v1/progress/:
    get:
      tags: [学习进度]
      summary: 获取全部进度
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Progress"

  # ========== 收藏 ==========
...